[Frontend] Radar widget refetch (#14448)
