+++ opencti-platform/opencti-front/src/components/dashboard/WidgetRadar.tsx
@@ -10,7 +10,7 @@ import { simpleNumberFormat } from '../../utils/Number';
 
 interface WidgetRadarProps {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  data: any[];
+  data: readonly any[];
   label: string;
   groupBy: string;
   onMounted?: OpenCTIChartProps['onMounted'];
