+++ opencti-platform/opencti-front/src/private/components/common/stix_core_objects/StixCoreObjectsRadar.tsx
@@ -1,0 +1,210 @@
+import React, { ReactNode, Suspense } from 'react';
+import { graphql, PreloadedQuery, usePreloadedQuery } from 'react-relay';
+import { useFormatter } from '../../../../components/i18n';
+import { buildFiltersAndOptionsForWidgets } from '../../../../utils/filters/filtersUtils';
+import WidgetContainer from '../../../../components/dashboard/WidgetContainer';
+import WidgetNoData from '../../../../components/dashboard/WidgetNoData';
+import WidgetRadar from '../../../../components/dashboard/WidgetRadar';
+import Loader, { LoaderVariant } from '../../../../components/Loader';
+import useDashboardViz from '../../../../components/dashboard/useDashboardViz';
+import WidgetNoHostEntity from '../../../../components/dashboard/WidgetNoHostEntity';
+import { Widget, WidgetDataSelection, WidgetHost } from '../../../../utils/widget/widget';
+import { StixCoreObjectsRadarDistributionQuery } from './__generated__/StixCoreObjectsRadarDistributionQuery.graphql';
+import { DashboardConfig } from '../../../../components/dashboard/dashboard-types';
+import { computeStartEndDates } from '../../../../components/dashboard/dashboard-viz-utils';
+
+const stixCoreObjectsRadarDistributionQuery = graphql`
+  query StixCoreObjectsRadarDistributionQuery(
+    $objectId: [String]
+    $relationship_type: [String]
+    $toTypes: [String]
+    $field: String!
+    $startDate: DateTime
+    $endDate: DateTime
+    $dateAttribute: String
+    $operation: StatsOperation!
+    $limit: Int
+    $order: String
+    $types: [String]
+    $filters: FilterGroup
+    $search: String
+  ) {
+    stixCoreObjectsDistribution(
+      objectId: $objectId
+      relationship_type: $relationship_type
+      toTypes: $toTypes
+      field: $field
+      startDate: $startDate
+      endDate: $endDate
+      dateAttribute: $dateAttribute
+      operation: $operation
+      limit: $limit
+      order: $order
+      types: $types
+      filters: $filters
+      search: $search
+    ) {
+      label
+      value
+      entity {
+        ... on BasicObject {
+          id
+          entity_type
+        }
+        ... on BasicRelationship {
+          id
+          entity_type
+        }
+        ... on StixObject {
+          representative {
+            main
+          }
+        }
+        ... on StixRelationship {
+          representative {
+            main
+          }
+        }
+        # use colors when available
+        ... on Label {
+          color
+        }
+        ... on MarkingDefinition {
+          x_opencti_color
+        }
+        # objects without representative
+        ... on Creator {
+          name
+        }
+        ... on Group {
+          name
+        }
+        ... on Status {
+          template {
+            name
+            color
+          }
+        }
+      }
+    }
+  }
+`;
+
+interface StixCoreObjectsRadarComponentProps {
+  queryRef: PreloadedQuery<StixCoreObjectsRadarDistributionQuery>;
+  dataSelection: Widget['dataSelection'];
+}
+
+const StixCoreObjectsRadarComponent = ({
+  queryRef,
+  dataSelection,
+}: StixCoreObjectsRadarComponentProps) => {
+  const { t_i18n } = useFormatter();
+  const data = usePreloadedQuery(
+    stixCoreObjectsRadarDistributionQuery,
+    queryRef,
+  );
+  const selection = dataSelection[0];
+  const distribution = data?.stixCoreObjectsDistribution ?? [];
+
+  if (distribution.length === 0) {
+    return <WidgetNoData />;
+  }
+  return (
+    <WidgetRadar
+      data={distribution}
+      label={selection.label ?? t_i18n('Number of entities')}
+      groupBy={selection.attribute ?? 'entity_type'}
+    />
+  );
+};
+
+interface StixCoreObjectsRadarProps {
+  variant?: string;
+  height?: number;
+  dataSelection: Widget['dataSelection'];
+  parameters?: {
+    title?: string;
+  };
+  popover?: ReactNode;
+  host?: WidgetHost;
+  config: DashboardConfig;
+  refreshRate?: number | null;
+}
+
+const DATA_SELECTION_TYPES = ['Stix-Core-Object'];
+
+const buildQueryVariables = (
+  resolvedDataSelection: WidgetDataSelection[],
+  config: DashboardConfig,
+): StixCoreObjectsRadarDistributionQuery['variables'] => {
+  const selection = resolvedDataSelection[0];
+  const dateAttribute = selection.date_attribute?.length
+    ? selection.date_attribute
+    : 'created_at';
+  const { startDate, endDate } = computeStartEndDates(config);
+  const { filters } = buildFiltersAndOptionsForWidgets(selection.filters, {
+    startDate,
+    endDate,
+    dateAttribute,
+  });
+
+  return {
+    types: DATA_SELECTION_TYPES,
+    field: selection.attribute ?? 'entity_type',
+    operation: 'count',
+    startDate,
+    endDate,
+    dateAttribute,
+    filters,
+    limit: selection.number ?? 10,
+  };
+};
+
+const StixCoreObjectsRadar = ({
+  variant,
+  height,
+  dataSelection,
+  parameters = {},
+  popover,
+  host,
+  config,
+  refreshRate = null,
+}: StixCoreObjectsRadarProps) => {
+  const { t_i18n } = useFormatter();
+  const { resolvedDataSelection, isMissingHostEntity, isPreviewMode, queryRef } = useDashboardViz<StixCoreObjectsRadarDistributionQuery>({
+    perspective: 'entities',
+    dataSelection,
+    host,
+    refreshRate,
+    query: stixCoreObjectsRadarDistributionQuery,
+    config,
+    buildQueryVariables,
+  });
+
+  if (isMissingHostEntity) {
+    return <WidgetNoHostEntity host={host} />;
+  }
+
+  return (
+    <WidgetContainer
+      padding="small"
+      height={height}
+      title={parameters.title ?? t_i18n('Distribution of entities')}
+      variant={variant}
+      action={popover}
+      showPreviewTag={isPreviewMode}
+    >
+      {queryRef && (
+        <Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
+          <StixCoreObjectsRadarComponent
+            queryRef={queryRef}
+            dataSelection={resolvedDataSelection}
+          />
+        </Suspense>
+      )}
+    </WidgetContainer>
+  );
+};
+
+export default StixCoreObjectsRadar;
